done refactoring

+++ py4phi/controller.py
@@ -13,21 +13,24 @@ from py4phi._encryption._pyspark_encryptor import _PySparkEncryptor
 
 from py4phi.logger_setup import logger
 from py4phi.consts import (
-    DEFAULT_CONFIG_NAME, DEFAULT_SECRET_NAME, DEFAULT_PY4PHI_OUTPUT_PATH
+    DEFAULT_CONFIG_NAME, DEFAULT_SECRET_NAME,
+    DEFAULT_PY4PHI_ENCRYPTED_PATH, DEFAULT_PY4PHI_DECRYPTED_PATH
 )
-
-handlers_mapping = {
-    DataFrame: PySparkDatasetHandler,
-
-}
 
 
 class Controller:
-    """Class to manipulate df."""
+    """Class to interact with py4phi components."""
+
+    HANDLERS_MAPPING = {
+        DataFrame: PySparkDatasetHandler,
+
+    }
 
     def __init__(self, dataset_handler: BaseDatasetHandler):
         self._dataset_handler = dataset_handler
         self.__encryptor = None
+        self._encrypted: bool = False
+        self._decrypted: bool = False
         self.__columns_data = None
         self._current_df = self._dataset_handler.df
         self._config_processor = ConfigProcessor()
@@ -40,35 +43,20 @@ class Controller:
 
         """
         logger.info('Printing active dataframe.')
-        self._current_df.show()
+        self._dataset_handler.print_df(self._current_df)
 
-    def encrypt_and_save(
+    def encrypt(
             self,
             columns_to_encrypt: list[str],
-            output_name: str = 'output_dataset',
-            save_location: str = DEFAULT_PY4PHI_OUTPUT_PATH,
-            config_file_name: str = DEFAULT_CONFIG_NAME,
-            encrypt_config: bool = True,
-            key_file_name: str = DEFAULT_SECRET_NAME,
-            **kwargs
-    ):
+    ) -> None:
         """
         Encrypt specified columns in dataset.
 
         Args:
         ----
         columns_to_encrypt (list[str]): List of columns to be encrypted.
-        output_name (str): Name of te output file.
-        encrypt_config (bool, optional): Whether to encrypt config.
-                                            Defaults to True.
-        save_location (str, optional): Folder location to save all the outputs.
-        config_file_name (str, optional): Name of config to be saved.
-                                    Defaults to DEFAULT_CONFIG_NAME.
-        key_file_name (str, optional): Name of config to be saved.
-                                Defaults to DEFAULT_SECRET_NAME.
-        kwargs (dict, optional): keyword arguments to be supplied to dataframe writing.
 
-        Returns: Encrypted dataframe.
+        Returns: None.
 
         """
         self.__encryptor = (
@@ -80,6 +68,40 @@ class Controller:
         logger.info(f'Kicking off encryption on current dataframe '
                     f'for columns: {columns_to_encrypt}.')
         self._current_df, self.__columns_data = self.__encryptor.encrypt()
+        self._encrypted = True
+        logger.info('Successfully encrypted dataframe.')
+
+    def save_encrypted(
+            self,
+            output_name: str = 'output_dataset',
+            save_location: str = DEFAULT_PY4PHI_ENCRYPTED_PATH,
+            config_file_name: str = DEFAULT_CONFIG_NAME,
+            encrypt_config: bool = True,
+            key_file_name: str = DEFAULT_SECRET_NAME,
+            **kwargs
+    ) -> None:
+        """
+        Save dataframe using provided location and parameters.
+
+        Args:
+        ----
+        output_name (str): Name of te output file.
+        encrypt_config (bool, optional): Whether to encrypt config.
+                                            Defaults to True.
+        save_location (str, optional): Folder location to save all the outputs.
+        config_file_name (str, optional): Name of config to be saved.
+                                    Defaults to DEFAULT_CONFIG_NAME.
+        key_file_name (str, optional): Name of config to be saved.
+                                Defaults to DEFAULT_SECRET_NAME.
+        kwargs (dict, optional): keyword arguments to be supplied to dataframe writing.
+
+        Returns: None.
+
+        """
+        if not self._encrypted:
+            logger.warn('No encryption action taken! '
+                        'Perhaps you forgot to encrypt your dataframe, aborting.')
+            exit(1)
 
         shutil.rmtree(save_location, ignore_errors=True)
         os.makedirs(save_location, exist_ok=True)
@@ -88,15 +110,55 @@ class Controller:
         self._config_processor.save_config(
             self.__columns_data,
             path=save_location,
-            conf_file_name=DEFAULT_CONFIG_NAME,
+            conf_file_name=config_file_name,
             encrypt_config=encrypt_config,
-            key_file_name=DEFAULT_SECRET_NAME
+            key_file_name=key_file_name
         )
         logger.debug(f'Saved config to: {save_location}.')
         self._dataset_handler.write(
             self._current_df,
             output_name,
-            DEFAULT_PY4PHI_OUTPUT_PATH,
+            save_location,
+            **kwargs
+        )
+        logger.info(f'Saved outputs to: {save_location}.')
+
+    def save_decrypted(
+            self,
+            output_name: str = 'output_dataset',
+            save_location: str = DEFAULT_PY4PHI_DECRYPTED_PATH,
+            **kwargs
+    ) -> None:
+        """
+        Save decrypted dataframe. Can be used without.
+
+        Args:
+        ----
+        output_name (str): Name of te output file.
+        encrypt_config (bool, optional): Whether to encrypt config.
+                                            Defaults to True.
+        save_location (str, optional): Folder location to save all the outputs.
+        config_file_name (str, optional): Name of config to be saved.
+                                    Defaults to DEFAULT_CONFIG_NAME.
+        key_file_name (str, optional): Name of config to be saved.
+                                Defaults to DEFAULT_SECRET_NAME.
+        kwargs (dict, optional): keyword arguments to be supplied to dataframe writing.
+
+        Returns: None.
+
+        """
+        if not self._decrypted:
+            logger.warn('No decryption action taken! '
+                        'Perhaps you forgot to decrypt your dataframe. '
+                        'Saving non-decrypted data.')
+        shutil.rmtree(save_location, ignore_errors=True)
+        os.makedirs(save_location, exist_ok=True)
+        logger.debug(f'Successfully prepared save location: {save_location}.')
+
+        self._dataset_handler.write(
+            self._current_df,
+            output_name,
+            save_location,
             **kwargs
         )
         logger.info(f'Saved outputs to: {save_location}.')
@@ -104,7 +166,7 @@ class Controller:
     def decrypt(
             self,
             columns_to_decrypt: list[str],
-            configs_path: str = DEFAULT_PY4PHI_OUTPUT_PATH,
+            configs_path: str = DEFAULT_PY4PHI_ENCRYPTED_PATH,
             config_file_name: str = DEFAULT_CONFIG_NAME,
             config_encrypted: bool = True,
             key_file_name: str = DEFAULT_SECRET_NAME
@@ -146,10 +208,11 @@ class Controller:
         )
 
         self._current_df = self.__encryptor.decrypt(decryption_dict)
+        self._decrypted = True
         logger.info('Successfully decrypted current df.')
 
 
-def init_from_path(
+def from_path(
         path: str,
         file_type: str,
         header: Optional[bool] = False,
@@ -171,15 +234,18 @@ def init_from_path(
     Returns: Controller object.
     """
     logger.setLevel(log_level)
+
     logger.debug("Initializing Dataset Handler")
     reader = PySparkDatasetHandler()
     logger.info(f"Reading dataframe using {type(reader)} "
                 f"from file: {path}, of type {file_type}.")
+
     reader.read_file(path=path, file_type=file_type, header=header)
+
     return Controller(reader)
 
 
-def init_from_dataframe(df, log_level: str | int = INFO) -> Controller:
+def from_dataframe(df, log_level: str | int = INFO) -> Controller:
     """
     Initialize controller object via given dataframe object.
 
@@ -193,11 +259,14 @@ def init_from_dataframe(df, log_level: str | int = INFO) -> Controller:
 
     """
     logger.setLevel(log_level)
-    if type(df) not in handlers_mapping.keys():
+
+    if type(df) not in Controller.HANDLERS_MAPPING.keys():
         raise ValueError(f'Unsupported object of type {type(df)}.')
-    handler = handlers_mapping[type(df)]()
+    handler = Controller.HANDLERS_MAPPING[type(df)]()
+
     logger.debug(f"Using {type(handler)} dataset handler")
     logger.info(f"Reading dataframe of type {type(df)}.")
+
     handler.read_dataframe(df)
 
     return Controller(handler)
